requirements.txt
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-docx==0.8.11
PyPDF2==3.0.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0